+++ src/main/java/com/optimisticchemicalmakers/mapfood/repositories/StoreRepository.java
@@ -17,7 +17,7 @@ public interface StoreRepository extends CrudRepository<Store, Long> {
     String HAVERSINE_PART = "(6371 * acos(cos(radians(:latitude)) * cos(radians(s.latitude)) * cos(radians(s.longitude) - radians(:longitude)) + sin(radians(:latitude)) * sin(radians(s.latitude))))";
 
     @Query(value="SELECT s.*, "+HAVERSINE_PART+" AS distance FROM store s HAVING distance <= :distance order by distance",nativeQuery = true)
-    List<Store> getNearestStores(@Param("latitude") Float latitude, @Param("longitude") Float longitude, @Param("distance") Long distance);
+    List<Store> getNearestStores(@Param("latitude") Double latitude, @Param("longitude") Double longitude, @Param("distance") Double distance);
  
 
 }
